% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{collapse_names}
\alias{collapse_names}
\title{Mutate author naming convention
`easyPubMed` returns complete names. This function preserves first initial and
last name.}
\usage{
collapse_names(df)
}
\arguments{
\item{df}{Data frame that is the output of `query_pm`.}
}
\description{
Mutate author naming convention
`easyPubMed` returns complete names. This function preserves first initial and
last name.
}
