% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{collapse_pm_query}
\alias{collapse_pm_query}
\title{Restructure Query
`easyPubMed` returns one row per author per publication. This function creates a
comma-separated list of authors, and one row per publication.}
\usage{
collapse_pm_query(df)
}
\arguments{
\item{df}{Data frame that is the output of `query_pm`.}
}
\description{
Restructure Query
`easyPubMed` returns one row per author per publication. This function creates a
comma-separated list of authors, and one row per publication.
}
