% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lw-functions.R
\name{get_first_author}
\alias{get_first_author}
\title{Get First Author
Munge author string to pull out only the first authors last name}
\usage{
get_first_author(author_strings)
}
\arguments{
\item{author_strings}{vector of authors from \code{query_pubmed()} output.}
}
\value{
vector of author first names
}
\description{
Get First Author
Munge author string to pull out only the first authors last name
}
