% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_synapse_table_rows}
\alias{get_synapse_table_rows}
\title{Get Synapse table}
\usage{
get_synapse_table_rows(synID, syn)
}
\arguments{
\item{synID}{The Synapse ID of a table to query from. Defaults to
"syn10242922"}

\item{syn}{Synapse client object}
}
\value{
Data frame of table contents
}
\description{
Get the contents of a Synapse table as a data frame and preserve the row IDs.
}
