% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lw-functions.R
\name{grant_query}
\alias{grant_query}
\title{Query Grant Serial Numbers}
\usage{
grant_query(grant_serial_nums, max = 99999)
}
\arguments{
\item{grant_serial_nums}{A vector of grant unique IDs.}

\item{max}{Number of results to return. Maxes out at 99999.}
}
\value{
A list of vectors of publication PubMed IDs for each grant serial number supplied.
}
\description{
Query Grant Serial Numbers
}
