% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-cleaning.R
\name{hacky_cleaning}
\alias{hacky_cleaning}
\title{Clean up funky text}
\usage{
hacky_cleaning(text)
}
\arguments{
\item{text}{String to update}
}
\value{
The text with replacement, if needed.
}
\description{
Removes html formatting, converts non-ascii characters to ascii, and does
several replacements:
Replaces '&amp;', amp;' with 'and'. Replaces '&#39;' with an apostrophe.
Replaces '&quot;' with single quote.
}
