% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_entity_name}
\alias{make_entity_name}
\title{Make Entity Names
Set first author, journal, year and Pubmed ID as the Synapse entity name.}
\usage{
make_entity_name(df)
}
\arguments{
\item{df}{Data frame from \code{"easyPubMed::article_to_df()"}.}
}
\description{
Make Entity Names
Set first author, journal, year and Pubmed ID as the Synapse entity name.
}
