% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{munge_pubmed}
\alias{munge_pubmed}
\title{Munge PubMed Query Output
PubMed query output still needs some finishing touches. Get date from pubdate, create entity names, etc.}
\usage{
munge_pubmed(dat)
}
\arguments{
\item{df}{metadata dataframe output from \code{query_pubmed}.}
}
\value{
cleaned results data frame
}
\description{
Munge PubMed Query Output
PubMed query output still needs some finishing touches. Get date from pubdate, create entity names, etc.
}
