% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{parse_summary_obj_list}
\alias{parse_summary_obj_list}
\title{Parse Summary Obj}
\usage{
parse_summary_obj_list(summary_obj_list)
}
\arguments{
\item{summary_obj_list}{list of esummary records output from \code{`entrez_summary()`}}
}
\value{
dataframe containing authors and doi for each summary object in the list.
}
\description{
Parse Summary Obj
}
