% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{pub_query}
\alias{pub_query}
\title{Query Publications PubMed IDs}
\usage{
pub_query(pub_pmids_list)
}
\arguments{
\item{pub_pmids}{A list of vectors of pubmed IDs.}
}
\value{
A dataframe where each column corresponds to a publication and contains useful metadata (pmid, doi, title, pubdate, journal, grant serial number)
}
\description{
Query Publications PubMed IDs
}
