% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{query_general}
\alias{query_general}
\title{Query for Pubmed IDs only}
\usage{
query_general(query, sleep_time = 0.35)
}
\arguments{
\item{query}{Character string with query. The query can be simple
(e.g. "alzheimer disease") or complex
(e.g. "((alzheimer disease[TIAB]) AND (cat[TW]) AND (dog[TW]))". See the
PubMed guide for search field descriptions and tags:
https://pubmed.ncbi.nlm.nih.gov/help/#search-tags.}

\item{sleep_time}{Amount of time, in seconds, to sleep between queries.
Without APIKey, Pubmed limits rate of queries to 3 per second. Default of
0.35 seconds.}
}
\value{
Vector of integer Pubmed IDs associated with the query.
}
\description{
Query for Pubmed IDs only
}
