% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{query_list_general}
\alias{query_list_general}
\title{Wrapper for Pubmed Query using general queries
This function wraps several functions:
- query Pubmed
- create an entity name from first author, journal, year and Id
- abbreviates the author names by first initial, last name
- creates one row per PubmedId
- query without associated PubmedIds will be left out of returned data
- query added as an additional column}
\usage{
query_list_general(queries, sleep_time = 0.35)
}
\arguments{
\item{queries}{A character vector of queries.}

\item{sleep_time}{Amount of time, in seconds, to sleep between queries.
Without APIKey, Pubmed limits rate of queries to 3 per second. Default of
0.35 seconds.}
}
\description{
Wrapper for Pubmed Query using general queries
This function wraps several functions:
- query Pubmed
- create an entity name from first author, journal, year and Id
- abbreviates the author names by first initial, last name
- creates one row per PubmedId
- query without associated PubmedIds will be left out of returned data
- query added as an additional column
}
