% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{query_list_pmids}
\alias{query_list_pmids}
\title{Wrapper for Pubmed Query
This function wraps several functions:
- query Pubmed
- create an entity name from first author, journal, year and Id
- abbreviates the author names by first initial, last name
- creates one row per PubmedId}
\usage{
query_list_pmids(vec)
}
\arguments{
\item{vec}{A character vector of PudmedIds.}
}
\description{
Wrapper for Pubmed Query
This function wraps several functions:
- query Pubmed
- create an entity name from first author, journal, year and Id
- abbreviates the author names by first initial, last name
- creates one row per PubmedId
}
