% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{query_pm}
\alias{query_pm}
\title{Query PubMed for publication data
Returns a data frame with one author per row. Error handling for
publications without an author list.}
\usage{
query_pm(pmid)
}
\arguments{
\item{pmid}{PubmedId as a string.}
}
\description{
Query PubMed for publication data
Returns a data frame with one author per row. Error handling for
publications without an author list.
}
