% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{query_pubmed}
\alias{query_pubmed}
\title{Query PubMed for Publication Metadata}
\usage{
query_pubmed(grant_serial_nums)
}
\arguments{
\item{grant_serial_nums}{A vector of grant unique IDs.}
}
\value{
A dataframe of all publications that correspond to provided grant serial numbers. Includes pmid, doi, title, abstract, year, journal, entity_name
}
\description{
Query PubMed for Publication Metadata
}
