% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-cleaning.R
\name{remove_unacceptable_characters}
\alias{remove_unacceptable_characters}
\title{Remove unacceptable characters}
\usage{
remove_unacceptable_characters(text)
}
\arguments{
\item{text}{String, or vector of strings, that might have
unacceptable characters.}
}
\value{
The text with replacement, if needed.
}
\description{
The primary goal for this function is to clean up entity names, which may
only contain: letters, numbers, spaces, underscores, hyphens, periods,
plus signs, apostrophes, and parentheses. Anything outside of this list is
considered an 'unacceptable character'. This will fix anything included in
[porTools::hacky_cleaning()].
Additionally, it will fix ':' (replaced with hyphen),
';' (replaced with period), '/' (replaced with space), ',' (removed),
square brackets (removed), and &amp;', '&', amp;' (all replaced with 'and').
}
