% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_up_multiannotations}
\alias{set_up_multiannotations}
\title{Format multi-value annotations
Reticulate and Python client can turn vectors into multiple value annotations.}
\usage{
set_up_multiannotations(df, value)
}
\arguments{
\item{df}{A data frame, the output of `query_list_pmids`}

\item{value}{The column name to parse for multiple value annotations.}
}
\description{
Format multi-value annotations
Reticulate and Python client can turn vectors into multiple value annotations.
}
