% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{store_as_annotations}
\alias{store_as_annotations}
\title{Store Pubmed Query output as annotations
This function is constrained by the `easyPubMed` output schema. DOI is stored as
the entity path; visualized as a hyperlink. The Synapse entity name is an
abbreviated entity name. emaining variables are non-specifically stored as
annotaitons.}
\usage{
store_as_annotations(parent, list)
}
\arguments{
\item{parent}{Synapse Folder Id that will store the entities.}

\item{list}{List of publications.}
}
\description{
Store Pubmed Query output as annotations
This function is constrained by the `easyPubMed` output schema. DOI is stored as
the entity path; visualized as a hyperlink. The Synapse entity name is an
abbreviated entity name. emaining variables are non-specifically stored as
annotaitons.
}
