% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{make_entity_name}
\alias{make_entity_name}
\title{Make Entity Names
Paste together first author, journal, year and Pubmed ID as the Synapse entity name.}
\usage{
make_entity_name(dat)
}
\arguments{
\item{dat}{metadata dataframe output from \code{query_pubmed()}.}
}
\value{
vector of entity names
}
\description{
Make Entity Names
Paste together first author, journal, year and Pubmed ID as the Synapse entity name.
}
